DjangoProject/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
Markdown==3.5.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
to==0.3
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
